Event processor refa
