+++ eu2020/common/event_processor.py
@@ -1,4 +1,6 @@
 import random
+from typing import Optional
+
 import eu2020.data.texts as texts
 import eu2020.common.utils as utils
 from eu2020.common.budget import Budget
@@ -12,6 +14,7 @@ class EventProcessor:
     p_events = set()
     flags = set()
     all_parties = Parties({})
+    all_events = list()
 
     def __init__(self, flags: set):
         self.flags = flags
@@ -22,6 +25,7 @@ class EventProcessor:
         for p in parties:
             if p not in self.all_parties.parties:
                 self.all_parties.parties[p] = parties[p]
+        self.all_events += p_events.events
 
     def add_stories(self, stories: list) -> None:
         for s in stories:
@@ -32,15 +36,37 @@ class EventProcessor:
                     break
             else:
                 raise ValueError(f"party not found: {s['party']}")
+        self.all_events += stories
+
+    def event_filter(self, ev: dict) -> bool:
+        if "condition" in ev:
+            if "flag" in ev["condition"]:
+                for c in ev["condition"]["flag"]:
+                    if c not in self.flags:
+                        return False
+
+            if "satisfaction" in ev["condition"]:
+                if ev["condition"]["satisfaction"]["op"] == "<":
+                    if self.all_parties.parties[ev["party"]]["satisfaction_pct"] >= ev["condition"]["satisfaction"]["value"]:
+                        return False
+                if ev["condition"]["satisfaction"]["op"] == ">":
+                    if self.all_parties.parties[ev["party"]]["satisfaction_pct"] <= ev["condition"]["satisfaction"]["value"]:
+                        return False
+        return ev["wait"] == 0
+
+    def get_event(self) -> Optional[dict]:
+        filtered = list(filter(self.event_filter, self.all_events))
+        if filtered:
+            return random.choice(filtered)
+        return None
 
     def process_events(self, budget: Budget) -> None:
-        for party_evs in self.p_events:
-            for _ in range(random.randint(0, 2)):
-                ev = party_evs.get_event()
-                if ev is not None:
-                    country = party_evs.get_parties().parties[ev["party"]]
-                    utils.print_text_in_box(country["name"])
-                    self.make_decision(ev, self.all_parties, budget)
+        for _ in range(random.randint(1, 3)):
+            ev = self.get_event()
+            if ev is not None:
+                country = self.all_parties.parties[ev["party"]]
+                utils.print_text_in_box(country["name"])
+                self.make_decision(ev, self.all_parties, budget)
 
     def make_decision(self, ev: dict, members: Parties, budget: Budget) -> None:
         print_log(ev["description"])
