+++ eu2020/common/parties_events.py
@@ -1,6 +1,3 @@
-import random
-from typing import Optional
-
 from eu2020.common.parties import Parties
 
 
@@ -20,31 +17,6 @@ class PartiesEvents:
     def add_event(self, ev: dict) -> None:
         ev["wait"] = 0
         self.events.append(ev)
-
-    def event_filter(self, ev: dict) -> bool:
-
-        if "condition" in ev:
-
-            if "flag" in ev["condition"]:
-                for c in ev["condition"]["flag"]:
-                    if c not in self.flags:
-                        return False
-
-            if "satisfaction" in ev["condition"]:
-                if ev["condition"]["satisfaction"]["op"] == "<":
-                    if self.parties.parties[ev["party"]]["satisfaction_pct"] >= ev["condition"]["satisfaction"]["value"]:
-                        return False
-                if ev["condition"]["satisfaction"]["op"] == ">":
-                    if self.parties.parties[ev["party"]]["satisfaction_pct"] <= ev["condition"]["satisfaction"]["value"]:
-                        return False
-
-        return ev["wait"] == 0
-
-    def get_event(self) -> Optional[dict]:
-        filtered = list(filter(self.event_filter, self.events))
-        if filtered:
-            return random.choice(filtered)
-        return None
 
     def get_parties(self) -> Parties:
         return self.parties
